requirements.txt
google-genai>=1.30.0
python-dotenv>=1.0.1
requests>=2.32.3
Pillow>=10.4.0
google-api-python-client>=2.166.0
google-auth>=2.35.0
